pyproject.toml
[tool.poetry]
name = "sf-dot"
version = "0.1.0"
description = ""
authors = ["Aditya Banerjee <aditya_banerjee@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
